+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## ??? (Not Released yet)
+
+* Mode maintenance campagnodon
+
 ## 2.0.6
 
 * Synchronisation des statuts: ajout d'un watch dog pour détecter les erreurs de synchronisation.